[oncology-pack] feat: - update pack mgmt and metadata  - implement Metastatic Site Linkage mapping to OMOP fact_relationship    - Upgraded fact_id_1/2 in DDL & DB to bigint to avoid surrogate key integer overflow - Extracted primary_condition_ref from FHIR condition-related extension URL - Project primary_condition_ref in condition resolve pass - Added Condition__fact_relationship edge to ETL plan - Seeded primary/metastatic relationship vocabulary concepts - Created Condition__fact_relationship Stage-2 join query for bidirectional mapping - Added and verified metastatic linkage test case in oncology pack - Updated run script arguments and pack documentation

+++ cases/_cdm_ddl.sql
@@ -395,9 +395,9 @@ CREATE TABLE cdm_ours_fhir.episode_event (
 
 CREATE TABLE cdm_ours_fhir.fact_relationship (
     domain_concept_id_1 integer NOT NULL,
-    fact_id_1 integer NOT NULL,
+    fact_id_1 bigint NOT NULL,
     domain_concept_id_2 integer NOT NULL,
-    fact_id_2 integer NOT NULL,
+    fact_id_2 bigint NOT NULL,
     relationship_concept_id integer NOT NULL
 );
